refactor: separated the widgets into separate classes instead of methods
