+++ lib/athlete/widgets/widgets.dart
@@ -3,5 +3,12 @@ export 'athlete_mint_approve_button.dart';
 export 'athlete_page_tooltip.dart';
 export 'athlete_sell_approve_button.dart';
 export 'build_long_chart.dart';
+export 'build_long_chart.dart';
+export 'build_page.dart';
+export 'build_short_chart.dart';
 export 'build_short_chart.dart';
 export 'buttons.dart';
+export 'graph_side.dart';
+export 'mobile_view.dart';
+export 'stats_side.dart';
+export 'web_view.dart';
